+++ galsim/integ.py
@@ -17,10 +17,9 @@
 #
 
 import numpy as np
-from functools import reduce
 
 from . import _galsim
-from .errors import GalSimError, GalSimRangeError, GalSimValueError, convert_cpp_errors
+from .errors import GalSimError, GalSimValueError, convert_cpp_errors
 
 def int1d(func, min, max, rel_err=1.e-6, abs_err=1.e-12):
     """Integrate a 1-dimensional function from min to max.
@@ -57,53 +56,6 @@ def int1d(func, min, max, rel_err=1.e-6, abs_err=1.e-12):
         return result
     else:
         raise GalSimError(result)
-
-def midpt(fvals, x):
-    """Midpoint rule for integration.
-
-    Parameters:
-        fvals:    Samples of the integrand
-        x:        Locations at which the integrand was sampled.
-
-    Returns:
-        midpoint rule approximation of the integral.
-    """
-    dx = [x[1]-x[0]]
-    dx.extend(0.5*(x[2:]-x[0:-2]))
-    dx.append(x[-1]-x[-2])
-    weighted_fvals = [w*f for w,f in zip(dx, fvals)]
-    return reduce(lambda y,z:y+z, weighted_fvals)
-
-def trapz(func, min, max, points=10000):
-    """Simple wrapper around 'numpy.trapz' to take function and limits as inputs.
-
-    Example usage::
-
-        >>> def func(x): return x**2
-        >>> galsim.integ.trapz(func, 0, 1)
-        0.33333333500033341
-        >>> galsim.integ.trapz(func, 0, 1, 1e6)
-        0.33333333333349996
-        >>> galsim.integ.trapz(func, 0, 1, np.linspace(0, 1, 1e3))
-        0.33333350033383402
-
-    Parameters:
-        func:       The function to be integrated.  y = func(x) should be valid.
-        min:        The lower end of the integration bounds.
-        max:        The upper end of the integration bounds.
-        points:     If integer, the number of points to sample the integrand. If array-like, then
-                    the points to sample the integrand at. [default: 1000].
-    """
-    if not np.isscalar(points):
-        if (np.max(points) > max) or (np.min(points) < min):
-            raise GalSimRangeError("Points outside of specified range", points, min, max)
-    elif int(points) != points:
-        raise TypeError("npoints must be integer type or array")
-    else:
-        points = np.linspace(min, max, points)
-
-    return np.trapz(func(points),points)
-
 
 def midptRule(f, xs):
     """Midpoint rule for integration.